Update nonlinear_auxiliary.R
The main change comes from the assignment of variables with T changed to "TRUE", not all = changed to <-, the main principle is: function parameter assignment with "=", variable assignment for "<-". "

In addition, for the definition of the matrix "byrow = TRUE" this way usually use "=", instead of the assignment symbol <-. This is a way of passing a function parameter without using the assignment "<-" symbol.